build-ai/requirements.txt
chromadb>=0.4.22
sentence-transformers>=2.2.2
openai>=1.12.0
requests>=2.31.0
python-dotenv>=1.0.0
tqdm>=4.66.0
PyYAML>=6.0
numpy>=1.24.0